db: nominal pakai bigint, fix: input u/ no ktp dan no debitur

+++ database/migrations/2024_01_22_031626_create_debiturs_table.php
@@ -22,8 +22,8 @@ return new class extends Migration
             $table->integer('kode_developer');
             $table->string('proyek_perumahan');
             $table->integer('kode_notaris');
-            $table->integer('plafon_kredit');
-            $table->integer('saldo_pokok');
+            $table->unsignedBigInteger('plafon_kredit');
+            $table->unsignedBigInteger('saldo_pokok');
             $table->string('alamat_agunan');
             $table->string('blok', 4);
             $table->integer('no');
